cleanup in example commands
